requirements.txt
python-dotenv~=1.0.1
discord-pretty-help>=2.0.7
discord.py>=2.4.0